+++ typings/module.d.ts
@@ -2,3 +2,4 @@ declare module '*.css' {
   const classes: { [key: string]: string };
   export default classes;
 }
+declare module '*.webp';