+++ packages/review-rating/CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.4](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@3.0.3...@heycar-uikit/review-rating@3.0.4) (2023-05-23)
+
+**Note:** Version bump only for package @heycar-uikit/review-rating
+
+
+
+
+
+## [3.0.3](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@3.0.2...@heycar-uikit/review-rating@3.0.3) (2023-05-09)
+
+
+### Bug Fixes
+
+* **header:** fix css bundle issue for header and review rating components ([#203](https://github.com/hey-car/heycar-uikit/issues/203)) ([7d04b6d](https://github.com/hey-car/heycar-uikit/commit/7d04b6dbceaa2c186dfde44de5a404ddfea5873d))
+
+
+
+
+
+## [3.0.2](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@3.0.1...@heycar-uikit/review-rating@3.0.2) (2023-05-02)
+
+
+### Bug Fixes
+
+* **review-rating:** fix CSS issue ([#198](https://github.com/hey-car/heycar-uikit/issues/198)) ([faff2ec](https://github.com/hey-car/heycar-uikit/commit/faff2ec9e73755ed235f2252a83763e84199ad35))
+
+
+
+
+
+## [3.0.1](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@3.0.0...@heycar-uikit/review-rating@3.0.1) (2023-04-28)
+
+
+### Bug Fixes
+
+* **review-rating:** gracefully handle no IntersectionObserver ([#196](https://github.com/hey-car/heycar-uikit/issues/196)) ([ea5d596](https://github.com/hey-car/heycar-uikit/commit/ea5d5969945b47dd6ab96ab1800de66761868c97))
+
+
+
+
+
+# [3.0.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@2.0.0...@heycar-uikit/review-rating@3.0.0) (2023-04-27)
+
+
+### chore
+
+* fix rollup css ([#194](https://github.com/hey-car/heycar-uikit/issues/194)) ([7d1fc92](https://github.com/hey-car/heycar-uikit/commit/7d1fc92711ce1ee6004359e39ecb46b8b6c49204))
+
+
+### BREAKING CHANGES
+
+* styles now need to be imported separately
+
+* build(rollup): update to fix css exports
+
+* build(package.json): updates to all packages to trigger version bump
+
+
+
+
+
+# [2.0.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/review-rating@1.1.0...@heycar-uikit/review-rating@2.0.0) (2023-04-26)
+
+
+### Build System
+
+* **bump versions:** bumping versions ([#193](https://github.com/hey-car/heycar-uikit/issues/193)) ([08bdb77](https://github.com/hey-car/heycar-uikit/commit/08bdb771e58c3bbecfd370bf3df8d3e2c7b490f2))
+
+
+### BREAKING CHANGES
+
+* **bump versions:** styles now need to be imported separately
+
+
+
+
+
 # 1.1.0 (2023-04-18)